Add tests for page.close, improve test spy
Record all calls to the client bridge in the global object `calls`.
This allows for in-order checks with mixed operation types.

Use Array.filter() to filter operations. The function by() creates a
filter function that checks for matches with the given properties.

Example:

    calls.filter( by( op: 'set', type: 'rwt.widgets.Shell' ) )

returns an array of all call operations for shells (maybe empty).