Maintain listeners in JavaScript
Instead of letting the client implementation keep a list of listeners,
keep them in the WidgetProxies in JavaScript code.

This allows us to simplify the client implementations and the
ClientBridge interface (only a single callback function will be
required, comparable to an OperationHandler). This will also ensure
that listener management behaves the same on all platforms.

Moreover, this change allows us to notify JavaScript listeners directly
from JS code, which also makes them easier to test.