Accept native bridge as parameter in Tabris._start
The method Tabris._start() accepts an implementation of the
NativeBridge interface as first parameter. This eliminates the
need for the global object "NativeBridge". "Tabris" is now the
only object in the global JavaScript namespace.

Throw an error if Tabris.create is called before Tabris is started,
i.e. there's no NativeBridge implementation.