Properly dispose children, rename destroy to dispose
Rename public `destroy` function to `dispose` to align with SWT and
reserve the term "destroy" for the internal destruction of a disposed
object.

Recursively destroy children of a disposed object to release the proxy
objects from memory. The client implementation is expected to destroy
children of disposed objects, so we have to cleanup the proxies as well.
To do so, proxy objects need to store parent and children.

Prevent method calls on disposed objects.