Create dedicated proxy for page
To keep the API simple, Tabris.createPage( properties ) creates both
a page and a composite to contain the page content.
Therefore, createPage() returns a proxy that delegates to the proxies
for page and composite. For example, the property "title" must be
delegated to the page, while "background" goes to the composite.

Create a dedicated constructor for page proxies. Add a _create method
to avoid side effects in the constructor and for better testability.

Change-Id: If9490193bfafd66426c0dfe16389d0f4e6c1dd66