Translate proxy references in properties to ids
The native bridge expects references to other widgets/objects as
id strings while the methods create and set support references to
the proxy objects directly.

Translate references to proxy objects to proxy ids.

This is not done recursively to avoid creating copies of objects and
arrays. Where references to other widgets are nested in properties,
these cases must be handled specifically. Currently, this is case for
layoutData, that can contain widget references in an array.

Fix #19

Change-Id: Ia8896d2f5330ae85b8a42bef109a04c95379c340