Enable using "tabris" as function to obtain proxies
The current workaround for accessing "service objects", i.e. singleton
objects with a pre-defined id, is to create a proxy instance:

    var geo = new tabris.Proxy( "tabris.Geolocation" );

This practice has the drawback that for every call, a new instance is
created. However, multiple proxies for the same id would not work as
expected, as each instance maintains its own listeners list.
When the tabris function is used to obtain a proxy, we can return the
same instance on every call:

    var geo = tabris( "tabris.Geolocation" );

Fix #13

Change-Id: I09693bb37f77e81e53766deda8dfd56a16ee314e