Destroy main shell on close event
The clients ask for confirmation on closing a Shell by sending a
"Close" event. The close is confirmed by destroying the Shell.

Fixes #28, but only for the main shell. I'd like to avoid supporting
the type Shell as this name is uncommon outside of SWT. We should
introduce a custom type "Dialog" instead, that is mapped to Shell.
This mapping would then also have to include the listener.

Change-Id: I4bcc32023bf0ed098554d96bde7723921fffdca3