Run tests on concatenated tabris.js
Run all *.spec.js files in test/js against the final tabris.js.
This ensures that
- the resulting tabris.js works as expected
- all specs are executed
- all source files are included

To do so, the tabris object must be reset properly in all tests (there
were test failures because of the different execution order). Create
a tabris._reset method.

Alternatively, we could introduce a constructor for the tabris object.
However, this would be a separate task.

Change-Id: I3151e2acc402df4f5b91071931a295c4a3295ebe