Add support for color strings
For ease of use and in preparation of Canvas support, accept colors
(properties `foreground` and `background`) as color strings in formats
known from web programming.

Get returns colors in rgba format.

At the same time, remove support for colors as arrays. The protocol
format for colors differs from the rgba function in the range of the
alpha value: the protocol expects alpha in the range 0..255, while it
has to be in the range 0..1 in the rgba function. Avoid confusion by
removing support for arrays.

Supported color formats:
- "#xxxxxx"
- "#xxx"
- "rgb(r, g, b)", where r, g, and b are numbers in range 0..255
- "rgba(r, g, b, a)", where a is a number in range 0..1
- basic color names from CSS3 spec.

Currently unsupported formats:
- rgb/rgba with percentages, e.g rgb(110%, 0%, 0%)
- hsl and hsla color functions
- extended color names
- `transparent` keyword

Change-Id: I0618967d68c95a8a8bc2d05b0412eeffbdf5a2fc