Workaround for Android back button issue
When UIProxy receives the `ShowPreviousPage` event from the client, it
calls `close()` on the Proxy for the actual page. However, this is not
the PageProxy, but a Proxy held in PageProxy._page, therefore it does
not have this method.

As a workaround, add a close method on the internal page proxy that
delegates to PageProxy.close.

See #38

Change-Id: I6a3931e5f47268525a2707f1bf3da530492bba9e