Provide drawing context for canvas
Support a graphics context on canvas widgets. The context supports a
subset of the HTML Canvas API [1].

For now, the drawing context can be obtained for a canvas by calling
`tabris.getContext(canvas, width, height)`. This is provional, we'll
have to come up with a decent API.

Just like in HTML, context fields such as lineWidth ignore invalid
values and issue a warning. We'll need to find a generic implementation
for parameter checks.

Includes a small demo for testing. Not all functions are equally
supported on both platforms yet.

[1] https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D

Change-Id: Iaceb29eb683a05f60331632bb5bade068178ee1e