Skip illegal property values
Passing illegal values to the native bridge will result in platform
specific error messages that are hard to understand for JS developers.

When we discover that a property value is illegal, we should therefore
not pass this value to the native bridge, but handle this problem
on the JS side.

Currently, only those values that must be translated to the protocol
format are checked by the encodeProperty() function, but we should
generalize value checks, at least in development mode.

For the sake of fault-tolerance, we should not cause the client to crash
but issue a warning and skip the property.

Change-Id: Ice4eac3e5b1a4784de461e577a08646e05793ba8