Introduce flush mechanism for canvas draw operations
JS code can only be executed when either tabris._start() or
tabris._notify() is called by the native side. Trigger a global
"flush" event at the end of both methods.

Render canvas drawing operations only in this event in order to reduce
the amount of draw operations in a canvas animation to a single call
per frame. This speeds up animations in V8 by about 10 .. 20%.

Change-Id: Ie45f79990241e3bc876edcb4a98e53f91a04417e