Add localStorage API
The storage API is based on the W3C specifications [1].
WebStorage.js contains the WebStorage constructor which is
instantiated as a window.localStorage object.

Storage event type implemented, but it does not get fired
as it should only be triggered on documents other than the
one whose actions fired the event.

Fixes #79

[1]: http://www.w3.org/TR/webstorage