Add jscs code style checker configuration
Our JS code should follow a consistent and commonly accepted code style.

JSCS [1] is a code style checker for JS that can be integrated with the
WebStorm IDE and Grunt build. This configuration is based on a preset
for the Google style quide [2], with some exceptions and additions.

I suggest that we evaluate this setting by trying to apply it to our
code base and enable the grunt checks when we're comfortable with it.

[1] https://www.npmjs.org/package/jscs
[2] https://google-styleguide.googlecode.com/svn/trunk/javascriptguide.xml

Change-Id: I89c4c6d310b4db97608e0d5c5245ef6af0949348