Handle (re-)setting a widget's parent
It's already possible to create widgets using the tabris.create()
method and include the property `parent`. When this property is set,
the widget is registered with the parent to enable proper disposal.

To allow creating widgets without a parent, handle the `parent`
property in `set` as well.

Let the proxy's `append` method also accept one or more proxy objects.

Handle the case that a page or a scroll composite is set as parent. In
this case, the inner composite of the page must be the parent.

Change-Id: I0d9288034598a8d45c2642ad788e247cb079dbf0