Create constructor and prototype for all widgets
Instead of manually creating Proxy subtypes, let the registerType
function create a constructor and a prototype that extends the Proxy
prototype and adds a set of additional members.

As an effect, every proxy instance has a constructor that reflects its
type (e.g. button instanceof tabris.Button) and an additional field
`type` that can be used in switch statements. This change also reduces
the amount of code needed to define a custom type.

When calling tabris.create() with an unregistered type, the type is
currently registered automatically. This is done to reduce the amount
of changes in this commit. Later, we should throw an error message when
trying to create an unregistered type.

Change-Id: I46fead9d896a834033918c024218dce083dffd0e