Let jasmine fail if specs couldn't be parsed
When running jasmine in the browser it can happen that all specs appear
to be are green, when in reality some just weren't executed due to
errors. By listening to errors on the window object we can check that no
specs were skipped while parsing them.

This isn't necessary when running jasmine from grunt, any uncaught
errors in phantom will cause the build to fail.

Change-Id: Id3bed1fe1a2ff641e0eaf6bfff84ff961c26f772