Remove support for proxy.append(type, properties)
Native objects should only be created using tabris.create(). This is
also a clear indication that those objects must be disposed by the
application (unless their parent is disposed).

Support proxy.append() only with proxies. Throw an error when this
method is called with anything else than a Proxy instance.

Change-Id: Ibc78b53d8a7b8305c807922ab28810e96bc9a830