Introduce dummy method measureText on canvas context
As long as this method is not implemented by the tabris clients,
provide a dummy to prevent crashes in existing code for HTML canvas.

According to the spec [1], this method should return an instance of
TextMetrics. This interface has only one method `width`, I assume it's
not worthwhile to implement a constructor.

[1] http://www.w3.org/html/wg/drafts/2dcontext/html5_canvas_CR/#dom-context-2d-measuretext

Change-Id: I2e181b267242989498d98df83399a82c751cd024