Update layout documentation
This document was initially meant as a spec for the clients. Turn it
into a more user-centric documentation.

Describe every single property. There is a little more redundancy, but
it should be easier to follow that an explanation of the generic
concepts and exceptions from them.

Add center and baseline properties (See #1, #64)

Intentionally omitted line breaks as some markdown implementations
(GFM) do not combine subsequent lines into one paragraph but turn
every newline into a <br> tag.

Change-Id: I7aaf1a14cb8ce8d4b124995422be366ae1b40159