Buffer operations to prepare batch calls
Since calls from JS to native are costly, we should reduce them by
aggregating operations and transfer them in chunks. This is supposed
to speed up animations with many operations in a single frame.

Introduce NativeBridge, which aggregates operations and transfers them
on flush. This is a preparation for passing all queued operations in a
single native call (#89).

The buffer is flushed whenever the JS context is left, i.e. on "flush"
events, to let the client render the changes. Moreover, to ensure
consistency, the buffered operations are flushed when the last
operation has a return value.

Change-Id: I585118ff2d29ac8c7c099ee17495c881bcb07aad