Merge subsequent set/create operations in native bridge
When a 'create' or 'set' operation is pending, the properties of
subsequent set calls for the same proxy are appended to the properties
of the pending operation.

With this implementation, we can call _setProperty for every single
property and have a single point for property checks and translations.

Moreover, appendTo() will no longer lead to an additional set operation
when called directly after the creation of a widget.

Change-Id: I461f36e6ba7ab5a38e8149349d052660650d3875