Do not send "listen" for unkown event types.
Listener are still attached, but a console.info log informs the
developer in case it's a typo. Do not use console.warn since it's still
a legitimate use case to trigger your own event types.

Events can be made known in the registerType function by giving objects
"_listen" and "_trigger" with entries for the local/local event name.
Either a function that is called or string that is used as the
translated event name, or just "true" if no translation is
required.

Change-Id: I8b133da486d4e7650eb149a727840520733b8e7f