Introduce CollectionView widget, remove List
CollectionView replaces the List widget. Same as List, it displays a
list of items (assigned to the "items" property), but instead of a
template, it defines a callback (assigned to "initializeCell") that
fills a cell with the widgets needed to display a data item. When a
data item is mapped to a cell, the cell receives a "itemchange" event.

See included snippet.

Change-Id: I3c07ce24483a0da43ce34390c7776a9facb63c5a