Add basic support for _checkProperty map in tabris.registerWidget
This is a first step towards full property checks. At this point there
is only a warning printed if the property has no entry in _checkProperty
that is truthy. (The actual _checkProperty maps are not part of this
commit.) Later we should also support functions to check the value and
not forward it to the client if the property or the given value is not
supported.

Change-Id: I0578e399c4988adb0d4478eb48c4cab370c3a6f3