Makes parallax demo finally fully work on iOS
On iOS we were able to move the titleComposite beyond our wished
top position. The problem would add up with subsequent events so 
that the titleComposite eventually disappeard. Reading the bounds
of that element only on change:bounds does not only work best on
both platforms, it also performs better because we don't have to
access the bounds in every scroll event.

Change-Id: I74b234a94e34bac1a070370c1a14af12efba4b43