Enable wrapping in Label by default
In mobile UIs, non-wrapping labels are hardly ever useful. When
intrinsic size is used, the text would not wrap anyway, and with a
fixed size, the text would be cut off.

So instead of mapping a create-only property "wrap" to the style flag
WRAP, I'd prefer to wait for a real use case and implement a runtime
property when needed.

Set style=["WRAP"] by default and remove it from all demos.

See #144

Change-Id: Ia7fb23bdba86a110b855501eb5d5321fd838ecec