Implement checks for children support
Throw an error when a widget is appended to a parent that does not
support children, or only supports children of another type. These are
severe problems in the application and we should not try to continue.

Support a new field "_supportsChildren" in registerType. This field
defaults to false, container widgets need to set it to true. It can
also be set to a function to check children selectively
(e.g. TabFolder only supports childs of type Tab).

Fix #129

Change-Id: Iaefb973a7a4ce39d6c2c09826e34fe86b4eaf642