Suppress warnings for internal create properties
For some widgets, the clients require certain properties to be set that
are not part of our API (e.g. style flags). These internal properties
are specified in registerType using the key _internalProperties.

Internal properties bypass encoding and property checks. Since
subsequent properties are merged in the native bridge, these properties
can still be overridden by user properties.

Rename _setNativeProperty to _setPropertyNative to reduce confusion.
The previous name seemed to suggest that there was a concept of "native
properties", as opposed to non-native ones.

Change-Id: Id24350e3032f0b386b76ed6a5365db28091787b5