Implement explicit support for property value checks
Property check functions are stored in tabris.PropertyChecks. Move
layoutData checks there and add image as an example. 

The checks must return the checked value, which gives them the option to
correct them. This is not to be used for protocol encoding, which should
be done in the setProperty methods if needed. If the check fails the
check function must throw an Error. The proxy then decides how to
proceed (currently the value is still given to the native bridge.)

One minor issue: Inspired by the layoutData checks I added console
warnings for images that have both a dimension and scale. However,
unlike with errors this currently won't let us print which widget caused
it. 

Change-Id: Id0e4f400981bffbee2143692347337797e072dd0