Extract encode/decode functions from Proxy.js
They are stored now in PropertyEncoding and added in
tabris.registerWidget, just like PropertyChecks.

Also, get rid of the "images" (plural) encoding, there is no property
with that name.

An argument could be made for merging the encoding and check functions,
since it's the same code in some cases and would get rid of the bloated
"_defaultSetProperty"/"_defaultGetProperty" maps. "_setProperty"/"_getProperty" would then only be needed for the rare case of name translation and target redirection. 

In any case this change is a step in the right direction since these
properties all belong to the widget type, not proxies in general.

Change-Id: I567fb71fa218da68091fe29178a09402e0bf80f8