Implement trivial selector API and make collection proxy-compatible
The only matcher syntax currently supported is for the widget type (what
whould be the tag name in XML/HTML). For now matcher strings can only be
given to the proxy.children method or collection.filter. Like with the
array.filter method it can also be a function.

This will allow tabris widgets to be handled in a fashion similar to how
jQuery handles HTML elements, except that "append" is missing from
collection.


Change-Id: Ic902cdfecbc0cffe1e9c7c20c9a02837bf4b3e97