Support "methods" field in the widget json definition files
Also add some methods to Widget.json as an example.

Not having the methods of a widget in the same overview as the
properties and events makes little sense, especially when we have
methods specific to certain widgets like Page and Canvas, and the "call"
method is not used or document at all. 

Change-Id: Ib954190a9dd1d6bc681e710a32c92d26868e1975