Implement appear and disappear events on Page
These events are fired when a page is about to become the active page
(`appear`) and when a page is no longer active (`disappear`). They can
be used to stop an animation when a page goes to background (see canvas
animation example).

The names "appear" and "disappear" do not directly relate to the concept
of an *active* page, however, "activate" and "deactivate" do not seem
appropriate. The appear / disappear events could later be generalized
for widgets that become visible / invisible.

Change-Id: I18e1e8b17821eab00a091367584953e32f06aa3c