Support search actions
Search actions are specialized Actions, that show a search field on
activation. The have an additional property `proposals` that accepts
a list of strings. In addition to the `selection` event, they fire
a `modify` event on typing in the search field, and a `submit` event
on submitting a search term.

Currently, SearchAction is an extended copy of Action. We plan to allow
to register types based on others (#236), so we'll get rid of this
redundancy.

No unit test, as SearchAction only adds plain declaration to Action.

Add a snippet for search actions. For the "action" snippet, use another
icon than search to avoid confusion.

Fix #227

Change-Id: I060b6f4c9b5f32f1750756e1fd23d25a29866e6a