Let "natural" property check normalize negative values
When a property that accepts natural numbers is set to a negative value,
instead of raising an error, set the property to `0`. This reduces the
need for checks in user code when using the result of a calculation.
It also aligns with the behavior of web APIs such as the rgb() function
in CSS.

Change-Id: I120a3b2281667a1e353dd6db30f1d1fbc3ef1aa1