Improve the reddit example
Simplify the design of the detail page by removing the semitransparent
label and use the page title for the image title.

Make the page background black in the detail page to be consistent with
the gallery demo for showing images.

This also prevents the overlapping of the details page label by the 
image introduced by the z-order change.

Change-Id: I27704ec76fae9515b83daa2492cf922ae32eec4e