Add internal methods modify items in CollectionView
We didn't yet agree on an API, but want to allow for experimentation
with dynamically inserting and removing items in a CollectionView.

Allow applications to insert and remove items using the internal
methods `_insert` and `_remove`. By overriding the method `_getItem`,
applications can trigger a reload when scrolling reaches the end of
the list.

Change-Id: I99de3b3a900d0806b9c212aeaa5a430aee5b7629