Implement support for built-in modules
Built-in modules are defined using Module.define. The "./" main module
can no longer be used as the root-level module since it is created after
the built-in modules. Instead the "module" module is used as the root
module.

Non-modules scripts currently have to rely on Module.require to request
built-in modules. One all scripts are defined as modules this method
will no longer be needed.

Convert global "util" object to native module "tabris-util"

Change-Id: I5eb51b542237e06b75def093562e60967c0d1345