Adjust examples and snippets to depend on tabris as a node module
A dependency to the nightly build tabris.tgz has been added to all
package.json files in snippets/examples. The "tabris" entry pointing to
the tabris.min.js has been removed. It is now possible (and necessary)
to install tabris via npm install. 

During the build the tabris module will not be installed from the tgz
url, as it would use the tabris version from the last nightly build.
Instead the tabris module is copied form the local build folder. 

Change-Id: I253ab750ad5db22fa4ea3ed4c6dfc20f712963a6