Support selector strings in layoutData
This will allow layouts to be defined in a separate file/module from the
application code.

This features works by not encoding layoutData immediately in
PropertyEncoding.js, but instead return a function that can either
return the layoutData as given (for the getter), or with the selectors
resolved to the widget cid. 

If a selector can not be resolved the function will throw an error and
the widget will attempt to give the layoutData to the native client with
the next flush. If the selector can not be resolved again, the
layoutData computation will be forced to continue, using 0 in place of
the unresolved selector. This mirrors the behavior of referencing
widgets that are not actually siblings of the layouted widget.

Change-Id: If6ce14a0d9792732a7051a3b93f3e45634c0bdbe