Remove DateTime from input example
It's confusing to showcase widgets that aren't supported by Tabris.js.
DateTime is the only remaining case of an unsupported widget. It's also
the only occurrence of the term "rwt" in our examples.

We should come up with some kind of DatePicker API and add this to our
examples.

Fix #282

Change-Id: I8121368691a4ca00699af7364fd88678555d5b64