Let the animate method a return an Animation instance
The Animation object can be used to receive start/progress/completion
events, or to cancel the animation. No start method. Once the animation
is complete or canceled the object is useless, and the internal proxy
will be disposed.

It is neccessary to have the complete code of the "on"/"off" methods
(currently shared between tabris.Events and tabris.Proxy) in
tabris.Events, as we need the _listen callback in the Animation instance
call the _proxy._listen method.

Change-Id: Iefc583a414f51863ff75872c369c129f316915d1