Ensure the find method returns no duplicates
Applying the find method of a collection that contains a parent as well
as it's children could previously result in duplicates. That is not only
unexpected but could also potentially impact performance. 

Since all widgets have a cid it's simple to use them to check that they
haven't been encountered before in this query.

Change-Id: Ia196172ae1b1f577d7f0e4c19cd314b964c4c89b