Allow the proxy set method to be used with unkown properties
Introduce Tabris.Properties. It's a mixin for other properties, similar
to Tabris.Events. It has the callbacks _applyProperty and _readProperty.
_applyProperty will be called before a property is set, and may return
true if the property shall be stored in the proxy. _readProperty will be
called in a "get" if there is currently no property of the given named
stored in the proxy.

Also remove try-catch around _setProperty. If a _setProperty function
throws, it is expected to be a bug. 

Change-Id: I9ef156431871cc72e0cd506d674d6f6b481a754b