Remove support for _properties: true
Setting _properties to true in the registerProxy function is an
undocumented feature that allows all properties that are set to be
forwarded to the native client. This is not a good practice, as it would
cause crashes instead of warnings if we make mistakes in our internal JS
code.

This will require updates to the cordova.plugin type, as it uses this
feature. Using this change with current clients will keep the
application from starting.

Change-Id: I391f0b914f05fa6e25eeb2e8883d3c2dead2a3b1