Introduce methods to modify CollectionView items
In order to support dynamic collection views, we need API for adding
and removing items. We can't track changes to the array in the `items`
property as there is no Array.observe on iOS.

Provide methods `insert` and `remove` that allow changing the items
array instead of replacing it. For dynamic collection views, these
methods can be used instead of setting the items property.

We may add support for trackable list types such as Backbone.Collection
later, in order to separate model operations from the widget, however
that's out of scope for now.

Fix #254

Change-Id: I5c6da36f45b687d254c30e465dce3c9995ee8f3c