Add dynamic layout demo
This demo changes layout depending on device orientation (or, more
precisely, the aspect ratio of the page). Since it's not a real-world
like example I added it to snippets, even if it's using multiple
modules. While in-lining the layout modules would be possible, it's 
not a pattern that we would want to recommend.

Change-Id: Icb23d98974b899e84fd14b6bfd75da62a2a34648