Introduce tabris.device object
Information provided by Tabris.js should be available in the `tabris`
namespace. Introduce an object `tabris.device` that acts as a proxy for
the native service object `tabris.Device`. It provides access to
read-only properties such `model`, `language`, etc.

Keep publishing the global `device` object, but only with those
properties that are also available from the Cordova device plug-in:
`model`, `platform`, and `version`.

Change-Id: I9352d30274281d90129166753dc74ba29880ada6