Change syntax for events in tabris.registerType/registerWidget
Instead of _listen and _trigger there is now only the _events map. Valid
values are true (no changes), a string (translates event name in both
directions) or a map with the optional keys "name" (string), "listen",
and "trigger" (both functions). 

There is a bit of compatibility code in registerType to support the
current cordova.js. 

A 1:n or n:n mapping of event names is not yet supported. Each public
event type can only be mapped to one native event type and vice versa.

Change-Id: Ibd8d8905cc20562a669d41e91a01a1a63f9a39b1