Fix cordova example error caused by badge plugin update
As of version 0.6.2 of the plugin the promptForPermission function has
been deprecated. The replacement is registerPermission, which requires a
callback. The old method delegates to the new one, and without the
callback causes a crash.

Sine registerPermission is now optional, we can just remove the line.

See https://github.com/katzer/cordova-plugin-badge#version-062-01032015

Change-Id: Ia136a01b1c2770d22f942c5b62765bb5048c319b