Make second parameter optional in setTimeout/setInterval
The second parameter `delay` can be left out in the Web API. Support
that in Tabris.js as well.

Moreover, the spec [1] demands that values must be converted to
non-negative integers following a series of specified transformations.
Instead of a full implementation of this conversion, handle only the
most obvious cases in order to keep the code fast and simple: convert
non-numeric or negative values to zero and round values.

Fix #305

[1] http://www.w3.org/TR/2011/WD-html5-20110525/timers.html#get-the-timeout

Change-Id: I01e54ecaf984329314e045f59d91e1d505d42ccb