Do not send listen operations for dispose event
With the latest revision of the event configuration options the syntax
that we used for the dispose event no longer has the desired effect.
(Not issuing listen operations or print any warnings.) Simply removing
it does.

I decided against writing a test, since it would only be testing that
the default behavior is true for a specific event type.

Change-Id: I12ebdee90cec0a1cfdf01f418c4af2e5d0c83ba8