Remove animation progress event
The event can not be supported in iOS, therefore it has to go.

Fix the tray demo to use 3 simultaneous animations instead (tray
translation, overlay opacity and icon rotation). 

Also some further refactoring on the tray demo to align iOS and Android
behavior. Since it's not possible to touch the tray during the 
animation in iOs, the tray demo also no longer supports this. Another
minor change is that when dragging beyond the maximm tray position and
back, the finger/tray offset stays consistent.

Fix #319

Change-Id: I75c4bc49bff7fbffa4ed1704645a4a3faaf5d233