Fix image property of Page
Getting the value of the "image" property of "page" was impossible due
to incomplete client support and broken decoding. It was also not
nullable.

Now the value is stored in JavaScript (no client support needed), and
supports null as a value (set and get). The decoding also does no longer
add unset width/height/scale properties to the image object.

Fix #333