Introduce page selector as convenience API for Drawer
With the support for custom drawer content, there will be no default
content in the navigation drawer anymore. In order to add a page
selector without a lot of code, provide the convenience method
tabris.createPageSelector().

This method creates a CollectionView with items for all top-level
pages. By default, this view's layoutData fill the entire parent.
This allows for a simple one-liner:

  tabris.create("Drawer").append(tabris.createPageSelector());

The default properties can be overwritten. For example, when custom
content has been added to the navigation drawer, the page selector can
be included at a custom position:

  tabris.createPageSelector({
    layoutData: {left: 0, top: 200, right: 0, bottom: 0}
  }).appendTo(drawer);

It's also possible to add pages selectively:

  tabris.createPageSelector({
    items: [page1, page2]
  }).appendTo(drawer);

Change-Id: I395c901a572b90bb3be7e834f5f1ccf986ec277e