Add method tabris._defineModule to allow defining "built-in" modules
This will allow us to provide modules not found on npm (or found only
under a different name), like those built into node. 

The module system has been changed slightly so that modules that are
defined that way can be initialized lazily. This ensures that the
depenencies they may have are not resolved before they are actually
needed.

Change-Id: I1b250fde717b6f97f7054eda643203a9c408c990