Change animate return value, fire animation events on widgets
The approach of adding animation listeners on the object returned by
the animate() method is flawed. Since this method starts an animation,
application cannot add a start listener before that. A quick animation
may even be completed before a completion listener can be added.

Remove the returned object introduced by commit 8561a2ee and let the
animate() method return the widget again to allow method chaining.

Fire events `animationstart` and `animationend` on the widget instance.
These event names are also used in the web animations API [1].

Make animation options available on the event in an `options` field.

Support additional option `name` that can be used to identify an
animation.

Fix #302

[1] http://dev.w3.org/csswg/css-animations/#event-animationevent

Change-Id: Iec234a4ae55fe1cbbfb597d65281abf4b9e5ec3e