Fix missing page composites issue on iOS
As a of commit deb7a094, the page Composite does not have a parent when
it is is to the Page's "control" property. The iOS client can't handle
this case.

We can't use the regular _setParent method before creating the _Page
as this would trigger an "addchild" event before the _Page is created.
The PageSelector relies on the properties from the _Page.

To fix the actual issue, set the page composite's parent before creating
the _Page and register it as a child afterwards.

Change-Id: I49d4a7b19e71fe949415914917de7b766286897b