Create GestureRecognizer only when adding a gesture listener
When providing pre-configured gestures (later on) we don't want to
pre-emptively create recognizer for all possible combinations. Therefore
we need a mechanism to create and dispose recognizer on adding/removing
a gesture  listener.

This implementation assumes that the listener is attached after the
gesture is configured. If a gesture is configured after the listener is
attached, no events will be issued.

Change-Id: I82a3b5d0f002858f6227e990fba09e14cd923f8d