Support "tap" and "longpress" as pre-configured gestures
It's very inconvinient having to configure even the most common gestures
before they are useable. This commits introducs the concept of
pre-configured gestures, for now just "tap" and "longpress". Other
gestures will follow.

Pre-configure gestures will not be removed when setting a gestures
object without them in it, but they can be overwritten if the gestures
object contains one.

Change-Id: I60c8635ac73a5611db3ed8059a3ca52406b10c3e