Dispose animation objects on widget disposal
The animation object must be disposed of when the corresponding widget
is disposed, otherwise it won't ever be garbage collected (it's kept in
tabris._proxies).

This also ensures that animation events will not be triggered on
disposed widgets. Add a test to ensure that _notify does not trigger
events on disposed widgets.

Fix #337

Change-Id: I330b03aed583c3dcee47a6ee14c9e4864824158b