Revise Getting Started Guide and related documents
Merge "Getting Started" and "Your First App" to reduce redundancy and
improve clarity. Move most of the section on the console object to
w3c-api.md.

Revise order, add screenshots, fix a few mistakes, add sections on
scratchpad and snippets, change/add a couple of sentences to prevent
confusion or misconceptions.

Add Quickstart guide for the impatient.

Link both to each other and in index.md. Can not link to either in the
readme.md since the tabrisjs.com login prevents that, so link to
tabrisjs.com itself.

Change-Id: I33c7deb0a41db30c30b9499b8e98178f2ea9ec1e