Implement _on and _off for private listeners
A listener registered with "_on" can only be removd with "_off". 
A listener registered with on "can" only be removed with "off".

Otherwise it's not relevant whether the listener was registered with
"_on" or "on".

This change allows us to add / remove framework-internal listeners 
without interfering with listeners added by application code.

Change-Id: Ie51ddf2647d29257ee1db370225b4fc7d335a37c