Enable property caching
Properties that have been set in JavaScript should be cached as JS
values so no native "get" call is required the next time they are read.
Use the "nocache" flag to exclude properties that can be changed on 
the native side (e.g. "selection") or that are already stored with a 
custom setter.

The initial value is still read by a native "get". This may change in
the future: Unless the initial value is platform dependent it can 
be hard-coded. This would further reduce "get" calls and fix a few 
properties where the native "get" is currently not supported by one 
or both of the platforms.

Change-Id: If83954361c2a70cf9b5eac13c488355b05ce7986