Normalize widget property configuration maps in registerType
This simpliefies the code a bit, but it also means that any kind of
post-registration manipulation on the _properties object (used in tests
and in the launcher app) must be done in the normalized format.

Also remove the feature of setting _properties: true  to allow any
property to be passed through to the native. It's no longer in use
anyhwere.

Change-Id: I5d3fff5e79a52b686ea15a3f7048b04faeb41c0b