Rename "selection" events to "select", revise parameters
"select" is more consistent with other event names and can not be
confused with the "selection" property of the same name. It's also
shorter. The widget is always given as the first parameter, an "options"
object as the last. For CollectionView the second parameter is the
selected item, and the options object contains the index.

The previous events are still available, but a warning is printed to the
console. To achieve this an event alias configuration is now
implemented. It will later also be used to trigger "select" in addition
to "change:selection".

Don't throw an exception on Events.trigger() if the widget was disposed.
This can happen when triggering more than one event sequentially.

Change-Id: Ia4c52b261aea4b1cfa22a58c4dbf9f286e5c0c31